+++ templates/parts/admin/content-starter-faq.php
@@ -5,7 +5,8 @@
  * @package    Michelle
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.0.1
  */
 
 // Exit if accessed directly.
@@ -35,7 +36,7 @@ if ( ! is_customize_preview() ) {
 <!-- /wp:columns -->
 
 <!-- wp:image {"align":"wide","id":999999,"sizeSlug":"large","linkDestination":"none"} -->
-<figure class="wp-block-image alignwide size-large"><img src="https://raw.githubusercontent.com/webmandesign/demo-content/img/WMDEMO__people--pexels-august-de-richelieu-4427816.jpg" alt="Image alt text" class="wp-image-999999"/></figure>
+<figure class="wp-block-image alignwide size-large"><img src="https://raw.githubusercontent.com/webmandesign/demo-content/img/SAMPLE/_people--pexels-august-de-richelieu-4427816.jpg" alt="Image alt text" class="wp-image-999999"/></figure>
 <!-- /wp:image --></div></div>
 <!-- /wp:group -->