+++ changelog.md
@@ -1,5 +1,21 @@
 # Michelle Changelog
 
+## 1.0.1, 20210325
+
+### Updated
+- Reintroducing `accessibility-ready` tag so the theme can undergo a11y review
+- Starter theme images
+
+### File updates
+	changelog.md
+	readme.txt
+	style.css
+	templates/parts/admin/content-starter-about.php
+	templates/parts/admin/content-starter-contact.php
+	templates/parts/admin/content-starter-faq.php
+	templates/parts/admin/content-starter-home.php
+	templates/parts/admin/content-starter-services.php
+
 ## 1.0.0, 20210322
 
 - Initial release.